fix: tighten packed config API and settlement state flow
Agent-Logs-Url: https://github.com/EkuboProtocol/evm-contracts/sessions/fd0c48e7-e51e-4640-8fea-2f118d6ab803

+++ src/interfaces/ITWAMMRecoverableLiquidations.sol
@@ -120,8 +120,8 @@ interface ITWAMMRecoverableLiquidations {
         address debtToken,
         uint64 poolFee,
         uint16 collateralFactorBps,
-        uint256 triggerHealthFactorX18,
-        uint256 cancelHealthFactorX18
+        uint64 triggerHealthFactorX18,
+        uint64 cancelHealthFactorX18
     ) external;
 
     function getPairConfig(address collateralToken, address debtToken, uint64 poolFee)
